Add EHS category pages and update navigation
