+++ app/ehs/page.tsx
@@ -1,28 +1,23 @@
 "use client";
 import Layout from "@/layout/Layout";
-import { List, ListItem, ListItemText, Typography, Paper } from "@mui/material";
+import { List, ListItem, ListItemText, Paper } from "@mui/material";
 import Link from "next/link";
 import EHSIntro from "@/components/EHSIntro";
 import DailySafetyReview from "@/lib/EHS/Structured_Daily_Safety_Checklist.json";
 import EquipmentSafetyCheck from "@/lib/EHS/Equipment_Inspection_Checklist.json";
-import CorrectiveActionRegister from "@/lib/EHS/Corrective_Action_Register.json";
-import InternalAuditChecklist from "@/lib/EHS/Internal_Audit_Checklist.json";
-import InternalAuditReport from "@/lib/EHS/Internal_Audit_Report.json";
 import ModifiedDutyLog from "@/lib/EHS/Modified_Duty_Log.json";
 import ReturnToWorkAgreement from "@/lib/EHS/Return_to_Work_Agreement.json";
 import SafetyObjectiveWorksheet from "@/lib/EHS/Safety_Objective_Worksheet.json";
-import ThirtyDayIncidentFollowUp from "@/lib/EHS/30dayIncidientFollowUp.json";
 
-const forms = [
-  { def: CorrectiveActionRegister, slug: "corrective-action-register" },
-  { def: InternalAuditChecklist, slug: "internal-audit-checklist" },
-  { def: InternalAuditReport, slug: "internal-audit-report" },
-  { def: ModifiedDutyLog, slug: "modified-duty-log" },
-  { def: ReturnToWorkAgreement, slug: "return-to-work-agreement" },
-  { def: SafetyObjectiveWorksheet, slug: "safety-objective-worksheet" },
-  { def: DailySafetyReview, slug: "daily-safety-review" },
-  { def: EquipmentSafetyCheck, slug: "equipment-safety-check" },
-  { def: ThirtyDayIncidentFollowUp, slug: "30day-incident-follow-up" },
+const links = [
+  { title: "Incident Reporting", href: "/ehs/incident-reporting" },
+  { title: "Return to Work", href: "/ehs/return-to-work" },
+  { title: "Safety Checklists", href: "/ehs/safety-checklists" },
+  { title: ModifiedDutyLog.title, href: "/forms/modified-duty-log" },
+  { title: ReturnToWorkAgreement.title, href: "/forms/return-to-work-agreement" },
+  { title: SafetyObjectiveWorksheet.title, href: "/forms/safety-objective-worksheet" },
+  { title: DailySafetyReview.title, href: "/forms/daily-safety-review" },
+  { title: EquipmentSafetyCheck.title, href: "/forms/equipment-safety-check" },
 ];
 
 export default function EHSPage() {
@@ -39,9 +34,9 @@ export default function EHSPage() {
           Below are the available EHS forms. Select one to begin.
         </Typography> */}
         <List>
-          {forms.map((f) => (
-            <ListItem key={f.slug} component={Link} href={`/forms/${f.slug}`}>
-              <ListItemText primary={f.def.title} />
+          {links.map((l) => (
+            <ListItem key={l.href} component={Link} href={l.href}>
+              <ListItemText primary={l.title} />
             </ListItem>
           ))}
         </List>
